chore: making stucture of dicovery page more cleaner
