+++ catalyst_voices/apps/voices/lib/pages/discovery/sections/how_it_works.dart
