+++ catalyst_voices/apps/voices/lib/pages/discovery/state_selectors/campaign_categories_state_selector.dart
@@ -1,4 +1,4 @@
-import 'package:catalyst_voices/pages/discovery/campaign_categories.dart';
+import 'package:catalyst_voices/pages/discovery/sections/campaign_categories.dart';
 import 'package:catalyst_voices/widgets/widgets.dart';
 import 'package:catalyst_voices_blocs/catalyst_voices_blocs.dart';
 import 'package:catalyst_voices_localization/catalyst_voices_localization.dart';
