+++ catalyst_voices/apps/voices/lib/pages/discovery/state_selectors/current_campaign_selector.dart
@@ -1,4 +1,4 @@
-import 'package:catalyst_voices/pages/discovery/current_campaign.dart';
+import 'package:catalyst_voices/pages/discovery/sections/current_campaign.dart';
 import 'package:catalyst_voices/widgets/widgets.dart';
 import 'package:catalyst_voices_blocs/catalyst_voices_blocs.dart';
 import 'package:catalyst_voices_localization/catalyst_voices_localization.dart';
